Add a new option to set the local IP address
This commit adds a new option to explicitly set the local IP address
instead of relying on autodetection. This is useful if you want to run
the application inside of a Docker container, which will have a
Docker-internal IP address instead of the external, device-visible one.

The new option is called `network.local-ip`.